+++ docs/source/changes.rst
@@ -6,6 +6,15 @@ chronological order. Releases follow `semantic versioning <https://semver.org/>`
 all releases are available on `PyPI <https://pypi.org/project/pytask>`_ and
 `Anaconda.org <https://anaconda.org/conda-forge/pytask>`_.
 
+0.2.0 - 2022-xx-xx
+------------------
+
+- :pull:`211` allows for flexible dependencies and products which can be any pytree of
+  native Python objects as supported by pybaum.
+- :pull:`227` implements ``task.kwargs`` as a new way for a task to hold parametrized
+  arguments. It also implements :class:`_pytask.models.CollectionMetadata` to carry
+  parametrized arguments to the task class.
+
 
 0.1.9 - 2022-02-23
 ------------------